Refactor scoring to use float, improve UI and plotting
Changed the session score field and related logic from int to float in db.py and throughout the app. Enhanced waveform plotting with envelope downsampling and improved time axis. Refactored and modernized the SpeechPracticeApp UI, including better button/icon handling, session history, and playback controls. Updated scoring logic in transcribe_worker.py to use a new scaling function for clarity-to-score mapping. Improved code structure and type annotations for maintainability.

+++ db.py
@@ -20,7 +20,7 @@ class PracticeSession(Base):
     transcript  = Column(String,  nullable=False)
     wer         = Column(Float,   nullable=False)
     clarity     = Column(Float,   nullable=False)
-    score       = Column(Integer, nullable=False)
+    score       = Column(Float, nullable=False)
 
 
 def get_engine(db_path: str = "sessions.db"):
@@ -56,7 +56,7 @@ def add_session(
     transcript: str,
     wer: float,
     clarity: float,
-    score: int,
+    score: float,
 ):
     ts = datetime.now().isoformat(timespec="seconds")
     sess = PracticeSession(
